python/bioinf_practice/requirements.txt
biopython>=1.78
numpy>=1.19.1